Allow to run the default media player (closes #57)
If the gsetting rundefault is set to true (default to false) and no player is
running clicking on the status icon will launch the default media player.

Obviously this feature is not available when the extension is integrated in
the volume menu.

The default media player can be set in the gnome-control-center (details panel).