Simplify the timer
Timeouts will repeat if the function they call doesn't return false. There's no need for _startTimer to call it's self over and over again to reset the timeout when you can call a function that returns this.trackTime += 1 (Non false) in the timeout thus making it repeat until it's removed with _stopTimer.