Don't hardcode the sort order of the playlists.
Currently the sort order in _getPlaylists is hardcoded to "Alphabetical".

This fixes that by checking the Orderings prop and using "Alphabetical" if "Alphabetical" is in the provided Orderings or if no Orderings are provided. Otherwise it uses the first sort order in the Orderings array if one is provided but "Alphabetical" is not in it.