Listen for the CanSeek prop change and respond accordingly.
If the CanSeek prop changes the extension should call this._getPosition to decide if it wants to show the position slider or not. A player may want to change this prop mid song if for example it's playing a http stream and it's buffering or hasn't finished it's initial buffering yet.