Dynamic menu width
Dynamically change the PanelIndicator and Aggregate Menu width based on the large cover size and the Aggregate Menu. The menu size is defined as the natural width of the Aggregate Menu or the large cover size + a small amount of padding, whichever is larger. The Aggregate Menu may grow in width depending upon theme and cover size to accommodate the cover but will return to it's default width when no active player is present and/or the extension is disabled.