widgets.js: Define classes using var keyword
As of GNOME 3.26 the gnome shell complains about classes that are defined using
the const keyword and which are at the same time exported to other modules.
To fix that we define all classes in widgets.js using the var keyword.

This is also what the GNOME devs did in the gnome-shell:
https://bugzilla.gnome.org/show_bug.cgi?id=785084

Signed-off-by: Alexander Rüedlinger <a.rueedlinger@gmail.com>